+++ core/app/models/spree/order.rb
@@ -488,13 +488,7 @@ module Spree
     def create_proposed_shipments
       adjustments.shipping.delete_all
       shipments.destroy_all
-
-      packages = Spree::Stock::Coordinator.new(self).packages
-      packages.each do |package|
-        shipments << package.to_shipment
-      end
-
-      shipments
+      self.shipments = Spree::Stock::Coordinator.new(self).shipments
     end
 
     def apply_free_shipping_promotions
@@ -665,5 +659,6 @@ module Spree
           break random_token unless self.class.exists?(guest_token: random_token)
         end
       end
+
   end
 end