+++ core/app/models/spree/stock/packer.rb
@@ -1,11 +1,11 @@
 module Spree
   module Stock
     class Packer
-      attr_reader :stock_location, :order, :splitters
+      attr_reader :stock_location, :inventory_units, :splitters
 
-      def initialize(stock_location, order, splitters=[Splitter::Base])
+      def initialize(stock_location, inventory_units, splitters=[Splitter::Base])
         @stock_location = stock_location
-        @order = order
+        @inventory_units = inventory_units
         @splitters = splitters
       end
 
@@ -18,17 +18,19 @@ module Spree
       end
 
       def default_package
-        package = Package.new(stock_location, order)
-        order.line_items.each do |line_item|
-          if line_item.should_track_inventory?
-            next unless stock_location.stock_item(line_item.variant)
+        package = Package.new(stock_location)
+        inventory_units.group_by(&:variant).each do |variant, variant_inventory_units|
+          units = variant_inventory_units.clone
+          if variant.should_track_inventory?
+            next unless stock_location.stock_item(variant)
 
-            on_hand, backordered = stock_location.fill_status(line_item.variant, line_item.quantity)
-            package.add line_item, on_hand, :on_hand if on_hand > 0
-            package.add line_item, backordered, :backordered if backordered > 0
+            on_hand, backordered = stock_location.fill_status(variant, units.count)
+            package.add_multiple units.slice!(0, on_hand), :on_hand if on_hand > 0
+            package.add_multiple units.slice!(0, backordered), :backordered if backordered > 0
           else
-            package.add line_item, line_item.quantity, :on_hand
+            package.add_multiple units
           end
+
         end
         package
       end