+++ core/app/models/spree/stock/splitter/base.rb
@@ -8,7 +8,7 @@ module Spree
           @packer = packer
           @next_splitter = next_splitter
         end
-        delegate :stock_location, :order, to: :packer
+        delegate :stock_location, to: :packer
 
         def split(packages)
           return_next(packages)
@@ -20,7 +20,7 @@ module Spree
         end
 
         def build_package(contents=[])
-          Spree::Stock::Package.new(stock_location, order, contents)
+          Spree::Stock::Package.new(stock_location, contents)
         end
       end
     end