+++ core/lib/spree/testing_support/factories/stock_factory.rb
@@ -12,19 +12,20 @@ FactoryGirl.define do
   factory :stock_package, class: Spree::Stock::Package do
     ignore do
       stock_location { build(:stock_location) }
-      order { create(:order_with_line_items, line_items_count: 2) }
-      contents []
+      contents       { [] }
+      variants_contents { {} }
     end
 
-    initialize_with { new(stock_location, order, contents) }
+    initialize_with { new(stock_location, contents) }
+
+    after(:build) do |package, evaluator|
+      evaluator.variants_contents.each do |variant, count|
+        package.add_multiple build_list(:inventory_unit, count, variant: variant)
+      end
+    end
 
     factory :stock_package_fulfilled do
-      after(:build) do |package, evaluator|
-        evaluator.order.line_items.reload
-        evaluator.order.line_items.each do |line_item|
-          package.add line_item, line_item.quantity, :on_hand
-        end
-      end
+      ignore { variants_contents { { build(:variant) => 2 } } }
     end
   end
 end