+++ core/spec/models/spree/calculator/shipping/flat_percent_item_total_spec.rb
@@ -10,14 +10,7 @@ module Spree
       let(:line_item2) { build(:line_item, variant: variant2) }
 
       let(:package) do
-        Stock::Package.new(
-          build(:stock_location),
-          mock_model(Order),
-          [
-            Stock::Package::ContentItem.new(line_item1, variant1, 2),
-            Stock::Package::ContentItem.new(line_item2, variant2, 1)
-          ]
-        )
+        build(:stock_package, variants_contents: { variant1 => 2, variant2 => 1 })
       end
 
       subject { FlatPercentItemTotal.new(:preferred_flat_percent => 10) }