+++ core/spec/models/spree/calculator/shipping/flexi_rate_spec.rb
@@ -6,18 +6,8 @@ module Spree
       let(:variant1) { build(:variant, :price => 10) }
       let(:variant2) { build(:variant, :price => 20) }
 
-      let(:line_item1) { build(:line_item, variant: variant1) }
-      let(:line_item2) { build(:line_item, variant: variant2) }
-
       let(:package) do
-        Stock::Package.new(
-          build(:stock_location),
-          mock_model(Order),
-          [
-            Stock::Package::ContentItem.new(line_item1, variant1, 4),
-            Stock::Package::ContentItem.new(line_item2, variant2, 6)
-          ]
-        )
+        build(:stock_package, variants_contents: { variant1 => 4, variant2 => 6 })
       end
 
       let(:subject) { FlexiRate.new }