+++ core/spec/models/spree/stock/package_spec.rb
@@ -4,27 +4,30 @@ module Spree
   module Stock
     describe Package do
       let(:variant) { build(:variant, weight: 25.0) }
-      let(:line_item) { build(:line_item, variant: variant) }
       let(:stock_location) { build(:stock_location) }
       let(:order) { build(:order) }
 
-      subject { Package.new(stock_location, order) }
+      subject { Package.new(stock_location) }
+
+      def build_inventory_unit
+        build(:inventory_unit, variant: variant)
+      end
 
       it 'calculates the weight of all the contents' do
-        subject.add line_item, 4
+        4.times { subject.add build_inventory_unit }
         subject.weight.should == 100.0
       end
 
       it 'filters by on_hand and backordered' do
-        subject.add line_item, 4, :on_hand
-        subject.add line_item, 3, :backordered
-        subject.on_hand.count.should eq 1
-        subject.backordered.count.should eq 1
+        4.times { subject.add build_inventory_unit }
+        3.times { subject.add build_inventory_unit, :backordered }
+        subject.on_hand.count.should eq 4
+        subject.backordered.count.should eq 3
       end
 
       it 'calculates the quantity by state' do
-        subject.add line_item, 4, :on_hand
-        subject.add line_item, 3, :backordered
+        4.times { subject.add build_inventory_unit }
+        3.times { subject.add build_inventory_unit, :backordered }
 
         subject.quantity.should eq 7
         subject.quantity(:on_hand).should eq 4
@@ -32,14 +35,16 @@ module Spree
       end
 
       it 'returns nil for content item not found' do
-        item = subject.find_item(variant, :on_hand)
+        unit = build_inventory_unit
+        item = subject.find_item(unit, :on_hand)
         item.should be_nil
       end
 
-      it 'finds content item for a variant' do
-        subject.add line_item, 4, :on_hand
-        item = subject.find_item(variant, :on_hand)
-        item.quantity.should eq 4
+      it 'finds content item for an inventory unit' do
+        unit = build_inventory_unit
+        subject.add unit
+        item = subject.find_item(unit, :on_hand)
+        item.quantity.should eq 1
       end
 
       # Contains regression test for #2804
@@ -53,61 +58,85 @@ module Spree
         variant1 = mock_model(Variant, shipping_category: category1)
         variant2 = mock_model(Variant, shipping_category: category2)
         variant3 = mock_model(Variant, shipping_category: nil)
-        contents = [Package::ContentItem.new(line_item, variant1, 1),
-                    Package::ContentItem.new(line_item, variant1, 1),
-                    Package::ContentItem.new(line_item, variant2, 1),
-                    Package::ContentItem.new(line_item, variant3, 1)]
+        contents = [ContentItem.new(build(:inventory_unit, variant: variant1)),
+                    ContentItem.new(build(:inventory_unit, variant: variant1)),
+                    ContentItem.new(build(:inventory_unit, variant: variant2)),
+                    ContentItem.new(build(:inventory_unit, variant: variant3))]
 
-        package = Package.new(stock_location, order, contents)
+        package = Package.new(stock_location, contents)
         package.shipping_methods.should == [method1]
       end
 
       it 'builds an empty list of shipping methods when no categories' do
         variant  = mock_model(Variant, shipping_category: nil)
-        contents = [Package::ContentItem.new(line_item, variant, 1)]
-        package  = Package.new(stock_location, order, contents)
+        contents = [ContentItem.new(build(:inventory_unit, variant: variant))]
+        package  = Package.new(stock_location, contents)
         package.shipping_methods.should be_empty
       end
 
       it "can convert to a shipment" do
-        subject.add line_item, 2, :on_hand, variant
-        subject.add line_item, 1, :backordered, variant
+        2.times { subject.add build_inventory_unit }
+        subject.add build_inventory_unit, :backordered
 
         shipping_method = build(:shipping_method)
         subject.shipping_rates = [ Spree::ShippingRate.new(shipping_method: shipping_method, cost: 10.00, selected: true) ]
 
         shipment = subject.to_shipment
-        expect(shipment.address).to eq subject.order.ship_address
-        expect(shipment.order).to eq subject.order
         expect(shipment.stock_location).to eq subject.stock_location
         expect(shipment.inventory_units.size).to eq 3
 
         first_unit = shipment.inventory_units.first
         expect(first_unit.variant).to eq variant
         expect(first_unit.state).to eq 'on_hand'
-        expect(first_unit.order).to eq subject.order
         expect(first_unit).to be_pending
 
         last_unit = shipment.inventory_units.last
         expect(last_unit.variant).to eq variant
         expect(last_unit.state).to eq 'backordered'
-        expect(last_unit.order).to eq subject.order
 
         expect(shipment.shipping_method).to eq shipping_method
       end
 
-      context "line item and variant don't refer same product" do
-        let(:other_variant) { build(:variant) }
+      it 'does not add an inventory unit to a package twice' do
+        # since inventory units currently don't have a quantity
+        unit = build_inventory_unit
+        subject.add unit
+        subject.add unit
+        expect(subject.quantity).to eq 1
+        expect(subject.contents.first.inventory_unit).to eq unit
+        expect(subject.contents.first.quantity).to eq 1
+      end
 
-        before { subject.add(line_item, 4, :on_hand, other_variant) }
-
-        it "cant find the item given wrong variant" do
-          expect(subject.find_item(variant, :on_hand)).to be_nil
+      describe "#add_multiple" do
+        it "adds multiple inventory units" do
+          expect { subject.add_multiple [build_inventory_unit, build_inventory_unit] }.to change { subject.quantity }.by(2)
         end
 
-        it "finds the item when given proper variant and line item" do
-          expect(subject.find_item(other_variant, :on_hand)).to eq subject.contents.last
-          expect(subject.find_item(other_variant, :on_hand, line_item)).to eq subject.contents.last
+        it "allows adding with a state" do
+          expect { subject.add_multiple [build_inventory_unit, build_inventory_unit], :backordered }.to change { subject.backordered.count }.by(2)
+        end
+
+        it "defaults to adding with the on hand state" do
+          expect { subject.add_multiple [build_inventory_unit, build_inventory_unit] }.to change { subject.on_hand.count }.by(2)
+        end
+      end
+
+      describe "#remove" do
+        let(:unit) { build_inventory_unit }
+        context "there is a content item for the inventory unit" do
+
+          before { subject.add unit }
+
+          it "removes that content item" do
+            expect { subject.remove(unit) }.to change { subject.quantity }.by(-1)
+            expect(subject.contents.map(&:inventory_unit)).not_to include unit
+          end
+        end
+
+        context "there is no content item for the inventory unit" do
+          it "doesn't change the set of content items" do
+            expect { subject.remove(unit) }.not_to change { subject.quantity }
+          end
         end
       end
     end