+++ core/spec/models/spree/stock/prioritizer_spec.rb
@@ -3,71 +3,81 @@ require 'spec_helper'
 module Spree
   module Stock
     describe Prioritizer do
-      let(:order) { create(:order_with_line_items, line_items_count: 2) }
+      let(:order) { mock_model(Order) }
       let(:stock_location) { build(:stock_location) }
+      let(:variant) { build(:variant) }
 
-      let(:line_item1) { order.line_items[0] }
-      let(:line_item2) { order.line_items[1] }
+      def inventory_units
+        @inventory_units ||= []
+      end
+
+      def build_inventory_unit
+        mock_model(InventoryUnit, variant: variant).tap do |unit|
+          inventory_units << unit
+        end
+      end
 
       def pack
-        package = Package.new(order, stock_location)
+        package = Package.new(order)
         yield(package) if block_given?
         package
       end
 
       it 'keeps a single package' do
         package1 = pack do |package|
-          package.add line_item1, 1, :on_hand
-          package.add line_item2, 1, :on_hand
+          package.add build_inventory_unit
+          package.add build_inventory_unit
         end
 
         packages = [package1]
-        prioritizer = Prioritizer.new(order, packages)
+        prioritizer = Prioritizer.new(inventory_units, packages)
         packages = prioritizer.prioritized_packages
         packages.size.should eq 1
       end
 
       it 'removes duplicate packages' do
         package1 = pack do |package|
-          package.add line_item1, 1, :on_hand
-          package.add line_item2, 1, :on_hand
+          package.add build_inventory_unit
+          package.add build_inventory_unit
         end
+
         package2 = pack do |package|
-          package.add line_item1, 1, :on_hand
-          package.add line_item2, 1, :on_hand
+          package.add inventory_units.first
+          package.add inventory_units.last
         end
 
         packages = [package1, package2]
-        prioritizer = Prioritizer.new(order, packages)
+        prioritizer = Prioritizer.new(inventory_units, packages)
         packages = prioritizer.prioritized_packages
         packages.size.should eq 1
       end
 
       it 'split over 2 packages' do
         package1 = pack do |package|
-          package.add line_item1, 1, :on_hand
+          package.add build_inventory_unit
         end
         package2 = pack do |package|
-          package.add line_item2, 1, :on_hand
+          package.add build_inventory_unit
         end
 
         packages = [package1, package2]
-        prioritizer = Prioritizer.new(order, packages)
+        prioritizer = Prioritizer.new(inventory_units, packages)
         packages = prioritizer.prioritized_packages
         packages.size.should eq 2
       end
 
       it '1st has some, 2nd has remaining' do
-        order.line_items[0].stub(:quantity => 5)
+        5.times { build_inventory_unit }
+
         package1 = pack do |package|
-          package.add line_item1, 2, :on_hand
+          2.times { |i| package.add inventory_units[i] }
         end
         package2 = pack do |package|
-          package.add line_item1, 5, :on_hand
+          5.times { |i| package.add inventory_units[i] }
         end
 
         packages = [package1, package2]
-        prioritizer = Prioritizer.new(order, packages)
+        prioritizer = Prioritizer.new(inventory_units, packages)
         packages = prioritizer.prioritized_packages
         packages.count.should eq 2
         packages[0].quantity.should eq 2
@@ -75,16 +85,17 @@ module Spree
       end
 
       it '1st has backorder, 2nd has some' do
-        order.line_items[0].stub(:quantity => 5)
+        5.times { build_inventory_unit }
+
         package1 = pack do |package|
-          package.add line_item1, 5, :backordered
+          5.times { |i| package.add inventory_units[i], :backordered }
         end
         package2 = pack do |package|
-          package.add line_item1, 2, :on_hand
+          2.times { |i| package.add inventory_units[i] }
         end
 
         packages = [package1, package2]
-        prioritizer = Prioritizer.new(order, packages)
+        prioritizer = Prioritizer.new(inventory_units, packages)
         packages = prioritizer.prioritized_packages
 
         packages[0].quantity(:backordered).should eq 3
@@ -92,20 +103,22 @@ module Spree
       end
 
       it '1st has backorder, 2nd has all' do
-        order.line_items[0].stub(:quantity => 5)
+        5.times { build_inventory_unit }
+
         package1 = pack do |package|
-          package.add line_item1, 3, :backordered
-          package.add line_item2, 1, :on_hand
+          3.times { |i| package.add inventory_units[i], :backordered }
         end
         package2 = pack do |package|
-          package.add line_item1, 5, :on_hand
+          5.times { |i| package.add inventory_units[i] }
         end
 
         packages = [package1, package2]
-        prioritizer = Prioritizer.new(order, packages)
+        prioritizer = Prioritizer.new(inventory_units, packages)
         packages = prioritizer.prioritized_packages
+        packages[0].should eq package2
+        packages[1].should be_nil
         packages[0].quantity(:backordered).should eq 0
-        packages[1].quantity(:on_hand).should eq 5
+        packages[0].quantity(:on_hand).should eq 5
       end
     end
   end